big d_a_alink pass (#214)
* d_a_alink pass1 wip

* more d_a_alink work

* remove asm

* more daalink work

* 20% dol code decompiled

* fixed some nonmatchings for building

* a few daalink functions and labeling some HIO data

+++ src/d/meter/d_meter2.cpp
@@ -324,10 +324,10 @@ int dMeter2_c::_create() {
 
     mMaxMagic = dComIfGs_getMaxMagic();
 
-    mNowOil = dComIfGs_getOil();
+    mNowOil = (s16)dComIfGs_getOil();
     dComIfGp_setItemNowOil(mNowOil);
 
-    mMaxOil = dComIfGs_getMaxOil();
+    mMaxOil = (s16)dComIfGs_getMaxOil();
 
     mNowOxygen = dComIfGp_getOxygen();
     dComIfGp_setNowOxygen(mNowOxygen);
